Add Script for Removing Unwanted System Apps.
Add init.d script to remove unwanted apps or apps that can be downloaded from the play store. Going with the philosophy if you can find the app on the Play Store then can choose to have it installed or not.  Otherwise just not that into Apollo or the CM file explorer.

Also started the beginnings of a general clean up script that will remove the above script after it has ran,